# docs/static/openapi/featuresignals.json
{
  "components": {
...
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "EvalContext": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "type": "object"
          },
          "key": {
            "description": "Target identifier",
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "type": "object"
      },
      "EvalResult": {
        "properties": {
          "duration_ms": {
            "type": "number"
          },
          "flag_key": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "rule_index": {
            "type": "integer"
          },
          "value": {}
        },
        "type": "object"
      },
...
      "EvaluateRequest": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/EvalContext"
          },
          "flag_key": {
            "type": "string"
          }
        },
        "required": [
          "flag_key",
          "context"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "Environment API key for server-side or client-side SDK access",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /v1/auth/login or /v1/auth/register",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "FeatureSignals",
      "url": "https://featuresignals.com"
    },
    "description": "Complete API reference for FeatureSignals \u2014 feature flag management platform.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "FeatureSignals API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/evaluate": {
      "post": {
        "description": "Evaluates a feature flag for the given context and returns its value, variation key, and reason.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalResult"
                }
              }
            },
            "description": "Evaluate a single flag \u2014 success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Evaluate a single flag",
        "tags": [
          "Evaluation"
        ]
      }
    },
...
    "/v1/agent/flag/{key}": {
      "get": {
        "description": "Get agent-readable flag details",
        "parameters": [
          {
            "description": "Flag key",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Project ID",
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "category": {
                      "type": "string"
                    },
                    "created_at": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "default_value": {},
                    "description": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "tags": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "type": {
                      "type": "string"
                    },
                    "updated_at": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Get agent flag \u2014 success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable entity"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get agent flag",
        "tags": [
          "Agent API"
        ]
      }
    },
    "/v1/agent/evaluate": {
      "post": {
        "description": "Evaluate a single flag with <5ms latency",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "attributes": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "environment": {
                    "type": "string"
                  },
                  "flag_key": {
                    "type": "string"
                  },
                  "key": {
                    "type": "string"
                  },
                  "project_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "flag_key",
                  "project_id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "eval_time_ms": {
                      "type": "number"
                    },
                    "flag_key": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    },
                    "value": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Evaluate flag \u2014 success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable entity"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Evaluate flag",
        "tags": [
          "Agent API"
        ]
      }
    },
    "/v1/agent/bulk-eval": {
      "post": {
        "description": "Evaluate up to 50 flags in bulk",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "attributes": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "environment": {
                    "type": "string"
                  },
                  "flag_keys": {
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 50,
                    "type": "array"
                  },
                  "key": {
                    "type": "string"
                  },
                  "project_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "project_id",
                  "flag_keys"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "results": {
                      "items": {
                        "properties": {
                          "flag_key": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          },
                          "value": {}
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "total_time_ms": {
                      "type": "number"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bulk evaluate \u2014 success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable entity"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Bulk evaluate flags",
        "tags": [
          "Agent API"
        ]
      }
    },
    "/v1/agent/create-flag": {
      "post": {
        "description": "Create a flag via agent API",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "default_value": {},
                  "description": {
                    "type": "string"
                  },
                  "key": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "project_id": {
                    "type": "string"
                  },
                  "tags": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "type": {
                    "type": "string"
                  }
                },
                "required": [
                  "key",
                  "name",
                  "project_id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "description": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "project_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Create flag \u2014 success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable entity"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create flag via agent",
        "tags": [
          "Agent API"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Production",
      "url": "https://api.featuresignals.com"
    }
  ],
...
}